+++ crates/acts/src/model/tests/info.rs
@@ -1,4 +1,5 @@
 use serde_json::Value;
+use std::sync::Arc;
 
 use crate::{
     ActRunAs, ModelInfo, NodeKind, PackageInfo, ProcInfo, TaskInfo, TaskState, Workflow, data,
@@ -71,7 +72,7 @@ fn model_info_task() {
     let workflow = Workflow::new();
     let node_data = NodeData {
         id: "nid".to_string(),
-        content: crate::scheduler::NodeContent::Workflow(workflow),
+        content: crate::scheduler::NodeContent::Workflow(Arc::new(workflow)),
         level: 0,
         parent_id: None,
         prev_id: None,
@@ -238,7 +239,7 @@ fn model_info_task_arr_to_value() {
     let workflow = Workflow::new();
     let node_data = NodeData {
         id: "nid".to_string(),
-        content: crate::scheduler::NodeContent::Workflow(workflow),
+        content: crate::scheduler::NodeContent::Workflow(Arc::new(workflow)),
         level: 0,
         parent_id: None,
         prev_id: None,